[dev] Replace Sink and Stream structs with SinkAndStream trait
Previously, there were separate structs for handling Sink and Stream operations in the Socket type. These were removed and replaced with a unified SinkAndStream trait. This both removes duplicated codes for handling different types of streams and also makes it easier to swap out different types of streams used within the Socket. In addition to this, error handling within the session.rs has been updated to use this new trait.